Add tests for loading the original InternVL2 checkpoints
Move the integration test to the new model folder and assert `config.image_token_id`
matches the checkpoint tokenizer's `<IMG_CONTEXT>` id.

That assertion is the one that matters: the model masks image positions with
`config.image_token_id` while the processor takes the id from the tokenizer, so if the
two disagree no image features are spliced in and the model answers as if no image was
given, without raising.
